[spi] Refactor smoketest
 - Move the implementation to a lib, so that it can be shered
 between several test with different flash partnumbers.
