+++ sw/device/tests/spi_host_flash_test_impl.h
@@ -1,0 +1,92 @@
+// Copyright lowRISC contributors.
+// Licensed under the Apache License, Version 2.0, see LICENSE for details.
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef OPENTITAN_SW_DEVICE_TESTS_SPI_HOST_FLASH_TEST_IMPL_H_
+#define OPENTITAN_SW_DEVICE_TESTS_SPI_HOST_FLASH_TEST_IMPL_H_
+
+#include "sw/device/lib/base/status.h"
+#include "sw/device/lib/dif/dif_spi_host.h"
+
+/**
+ * Send the sw_reset command.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_software_reset(dif_spi_host_t *spi);
+
+/**
+ * Read the sfdp header and check the signature.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_read_sfdp(dif_spi_host_t *spi);
+
+/**
+ * Send the sector erase command, read a page that it was erased.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_sector_erase(dif_spi_host_t *spi);
+
+/**
+ * Send the enable quad mode command.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_enable_quad_mode(dif_spi_host_t *spi);
+
+/**
+ * Program a complete page and read it back to verify.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_page_program(dif_spi_host_t *spi);
+
+/**
+ * Perform a fast read operation.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_fast_read(dif_spi_host_t *spi);
+
+/**
+ * Perform a dual read operation.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_dual_read(dif_spi_host_t *spi);
+
+/**
+ * Perform a quad read operation.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_quad_read(dif_spi_host_t *spi);
+
+/**
+ * Check if the flash supports 4-byte addressing.
+ *
+ * @param spi A spi host handler.
+ * @return True if 4-byte addressing mode is supported.
+ */
+bool is_4_bytes_address_mode_supported(void);
+
+/**
+ * Enable 4-byte addressing mode, write and read to a page that needs 4 byes to
+ * be addressed and disables the the 4-bytes addressing mode.
+ *
+ * @param spi A spi host handler.
+ * @return status_t containing either OK or an error.
+ */
+status_t test_4bytes_address(dif_spi_host_t *spi);
+
+#endif  // OPENTITAN_SW_DEVICE_TESTS_SPI_HOST_FLASH_TEST_IMPL_H_
