+++ sw/device/tests/spi_host_smoketest.c
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 #include <assert.h>
 
+#include "spi_host_flash_test_impl.h"
 #include "sw/device/lib/arch/device.h"
 #include "sw/device/lib/base/macros.h"
 #include "sw/device/lib/base/memory.h"
@@ -22,163 +23,6 @@ static_assert(__BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__,
               "This test assumes the target platform is little endian.");
 
 OTTF_DEFINE_TEST_CONFIG();
-
-// A data pattern to program into the chip:
-// From: http://www.abrahamlincolnonline.org/lincoln/speeches/gettysburg.htm
-static const char kGettysburgPrelude[256] =
-    "Four score and seven years ago our fathers brought forth on this "
-    "continent, a new nation, conceived in Liberty, and dedicated to the "
-    "proposition that all men are created equal.";
-
-// The SFDP structure we'll read out of the chip.
-typedef struct sfdp {
-  union {
-    struct {
-      spi_flash_testutils_sfdp_header_t header;
-      spi_flash_testutils_parameter_header_t param;
-    };
-    uint8_t data[256];
-  };
-  uint32_t *bfpt;
-} sfdp_t;
-
-sfdp_t sfdp;
-
-status_t test_software_reset(dif_spi_host_t *spi) {
-  // The software reset sequence is two transactions: RSTEN followed by RST.
-  dif_spi_host_segment_t op = {
-      .type = kDifSpiHostSegmentTypeOpcode,
-      .opcode = kSpiDeviceFlashOpResetEnable,
-  };
-  TRY(dif_spi_host_transaction(spi, /*cs_id=*/0, &op, 1));
-
-  op.opcode = kSpiDeviceFlashOpResetEnable;
-  TRY(dif_spi_host_transaction(spi, /*cs_id=*/0, &op, 1));
-  return OK_STATUS();
-}
-
-status_t test_read_sfdp(dif_spi_host_t *spi) {
-  TRY(spi_flash_testutils_read_sfdp(spi, 0, sfdp.data, sizeof(sfdp.data)));
-  LOG_INFO("SFDP signature is 0x%08x", sfdp.header.signature);
-  CHECK(sfdp.header.signature == kSfdpSignature,
-        "Expected to find the SFDP signature!");
-
-  uint32_t bfpt_offset = (uint32_t)sfdp.param.table_pointer[0] |
-                         (uint32_t)(sfdp.param.table_pointer[1] << 8) |
-                         (uint32_t)(sfdp.param.table_pointer[2] << 16);
-  sfdp.bfpt = (uint32_t *)(sfdp.data + bfpt_offset);
-  return OK_STATUS();
-}
-
-status_t test_erase(dif_spi_host_t *spi) {
-  TRY(spi_flash_testutils_erase_sector(spi, 0, false));
-
-  // Check that the first page of flash actually got erased.
-  uint8_t buf[256] = {0};
-  TRY(spi_flash_testutils_read_op(spi, kSpiDeviceFlashOpReadNormal, buf,
-                                  sizeof(buf),
-                                  /*address=*/0,
-                                  /*addr_is_4b=*/false,
-                                  /*width=*/1,
-                                  /*dummy=*/0));
-  uint8_t expected[256];
-  memset(expected, 0xFF, sizeof(expected));
-  TRY_CHECK_ARRAYS_EQ(buf, expected, ARRAYSIZE(expected));
-  return OK_STATUS();
-}
-
-status_t test_enable_quad_mode(dif_spi_host_t *spi) {
-  if (sfdp.param.length < 14) {
-    return INVALID_ARGUMENT();
-  }
-  uint8_t mech =
-      (uint8_t)bitfield_field32_read(sfdp.bfpt[14], SPI_FLASH_QUAD_ENABLE);
-  LOG_INFO("Setting the EEPROM's QE bit via mechanism %d", mech);
-  TRY(spi_flash_testutils_quad_enable(spi, mech, /*enabled=*/true));
-  return OK_STATUS();
-}
-
-// Program a pattern into the flash part and read it back.
-status_t test_page_program(dif_spi_host_t *spi) {
-  TRY(spi_flash_testutils_program_page(spi, kGettysburgPrelude,
-                                       sizeof(kGettysburgPrelude),
-                                       /*address=*/0, /*addr_is_4b=*/0));
-
-  uint8_t buf[256];
-  TRY(spi_flash_testutils_read_op(spi, kSpiDeviceFlashOpReadNormal, buf,
-                                  sizeof(buf), 0,
-                                  /*addr_is_4b=*/false,
-                                  /*width=*/1,
-                                  /*dummy=*/0));
-  TRY_CHECK_ARRAYS_EQ(buf, kGettysburgPrelude, ARRAYSIZE(kGettysburgPrelude));
-  return OK_STATUS();
-}
-
-// Read the flash device using the "fast read" opcode.
-status_t test_fast_read(dif_spi_host_t *spi) {
-  uint8_t buf[256];
-  TRY(spi_flash_testutils_read_op(spi, kSpiDeviceFlashOpReadFast, buf,
-                                  sizeof(buf), 0,
-                                  /*addr_is_4b=*/false,
-                                  /*width=*/1,
-                                  /*dummy=*/8));
-  TRY_CHECK_ARRAYS_EQ(buf, kGettysburgPrelude, ARRAYSIZE(kGettysburgPrelude));
-  return OK_STATUS();
-}
-
-// Read the flash device using the "fast dual read" opcode.
-status_t test_dual_read(dif_spi_host_t *spi) {
-  uint8_t buf[256];
-  TRY(spi_flash_testutils_read_op(spi, kSpiDeviceFlashOpReadDual, buf,
-                                  sizeof(buf), 0,
-                                  /*addr_is_4b=*/false,
-                                  /*width=*/2,
-                                  /*dummy=*/8));
-  TRY_CHECK_ARRAYS_EQ(buf, kGettysburgPrelude, ARRAYSIZE(kGettysburgPrelude));
-  return OK_STATUS();
-}
-
-// Read the flash device using the "fast quad read" opcode.
-status_t test_quad_read(dif_spi_host_t *spi) {
-  uint8_t buf[256];
-  TRY(spi_flash_testutils_read_op(spi, kSpiDeviceFlashOpReadQuad, buf,
-                                  sizeof(buf), 0,
-                                  /*addr_is_4b=*/false,
-                                  /*width=*/4,
-                                  /*dummy=*/8));
-  TRY_CHECK_ARRAYS_EQ(buf, kGettysburgPrelude, ARRAYSIZE(kGettysburgPrelude));
-  return OK_STATUS();
-}
-
-static bool is_4_bytes_address_mode_supported(void) {
-  enum { kSupportOnly3Bytes, kSupport3and4Bytes, kSupportOnly4Bytes };
-  uint32_t address_mode =
-      bitfield_field32_read(sfdp.bfpt[0], SPI_FLASH_ADDRESS_MODE);
-  return (address_mode == kSupport3and4Bytes ||
-          address_mode == kSupportOnly4Bytes);
-}
-
-status_t test_4bytes_address(dif_spi_host_t *spi) {
-  enum { kAddress = 0x01000100, kSectorSize = 4096 };
-  static_assert(kAddress % kSectorSize,
-                "Should be at the beginning of the sector.");
-
-  TRY(spi_flash_testutils_enter_4byte_address_mode(spi));
-  TRY(spi_flash_testutils_erase_sector(spi, kAddress, true));
-
-  TRY(spi_flash_testutils_program_page(spi, kGettysburgPrelude,
-                                       sizeof(kGettysburgPrelude), kAddress,
-                                       /*addr_is_4b=*/true));
-
-  uint8_t buf[256];
-  TRY(spi_flash_testutils_read_op(spi, kSpiDeviceFlashOpReadNormal, buf,
-                                  sizeof(buf), kAddress,
-                                  /*addr_is_4b=*/true,
-                                  /*width=*/1,
-                                  /*dummy=*/0));
-  TRY_CHECK_ARRAYS_EQ(buf, kGettysburgPrelude, ARRAYSIZE(kGettysburgPrelude));
-  return spi_flash_testutils_exit_4byte_address_mode(spi);
-}
 
 bool test_main(void) {
   dif_spi_host_t spi_host;
@@ -200,7 +44,7 @@ bool test_main(void) {
   status_t result = OK_STATUS();
   EXECUTE_TEST(result, test_software_reset, &spi_host);
   EXECUTE_TEST(result, test_read_sfdp, &spi_host);
-  EXECUTE_TEST(result, test_erase, &spi_host);
+  EXECUTE_TEST(result, test_sector_erase, &spi_host);
   EXECUTE_TEST(result, test_enable_quad_mode, &spi_host);
   EXECUTE_TEST(result, test_page_program, &spi_host);
   if (is_4_bytes_address_mode_supported()) {
